Add support for server products PR in isolation
